requirements.txt
python-telegram-bot[webhooks]
openai
python-dotenv
openai==0.28
git+https://github.com/openai/whisper.git
whisper
sqlalchemy
psycopg2-binary
llama-index